refactor : 스키마 정규화 + json 형식 변경
